% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cute_scatterplot.R
\name{cute_scatterplot}
\alias{cute_scatterplot}
\title{Title
Super dooper cute scaddater plot}
\usage{
cute_scatterplot(data, var1, var2, lmline = FALSE)
}
\arguments{
\item{data}{Input data}

\item{var1}{Numeric var 1}

\item{var2}{Numeric var 2}

\item{lmline}{TRUE if you want a fitted line; default is FALSE}
}
\value{
scatterplot
}
\description{
Title
Super dooper cute scaddater plot
}
\examples{
No example 4 u

}
